release: v1.18.1 — DNS sudoers wildcards hotfix (#272)
- Bump version.go to 1.18.1.
- CHANGELOG entry under [1.18.1] documenting the fix and the heal path
  for existing installs.
- Move the assertNoWildcardArgs Linux-renderer test to a //go:build
  linux file so the macOS test binary doesn't try to compile a call to
  renderLinuxSudoers (caught by CI on PR #272).
